+++ src/RcppExports.cpp
@@ -671,19 +671,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// cqp_subcorpus_query
-SEXP cqp_subcorpus_query(SEXP subcorpus, SEXP name, SEXP query);
-RcppExport SEXP _RcppCWB_cqp_subcorpus_query(SEXP subcorpusSEXP, SEXP nameSEXP, SEXP querySEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< SEXP >::type subcorpus(subcorpusSEXP);
-    Rcpp::traits::input_parameter< SEXP >::type name(nameSEXP);
-    Rcpp::traits::input_parameter< SEXP >::type query(querySEXP);
-    rcpp_result_gen = Rcpp::wrap(cqp_subcorpus_query(subcorpus, name, query));
-    return rcpp_result_gen;
-END_RCPP
-}
 // cqp_subcorpus_size
 int cqp_subcorpus_size(SEXP scorpus);
 RcppExport SEXP _RcppCWB_cqp_subcorpus_size(SEXP scorpusSEXP) {
@@ -759,6 +746,19 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< SEXP >::type corpus(corpusSEXP);
     Rcpp::traits::input_parameter< SEXP >::type registry(registrySEXP);
     rcpp_result_gen = Rcpp::wrap(cqp_load_corpus(corpus, registry));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cqp_subcorpus
+SEXP cqp_subcorpus(SEXP name, SEXP corpus, Rcpp::IntegerMatrix region_matrix);
+RcppExport SEXP _RcppCWB_cqp_subcorpus(SEXP nameSEXP, SEXP corpusSEXP, SEXP region_matrixSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type name(nameSEXP);
+    Rcpp::traits::input_parameter< SEXP >::type corpus(corpusSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerMatrix >::type region_matrix(region_matrixSEXP);
+    rcpp_result_gen = Rcpp::wrap(cqp_subcorpus(name, corpus, region_matrix));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -876,7 +876,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_RcppCWB_cqp_set_registry", (DL_FUNC) &_RcppCWB_cqp_set_registry, 1},
     {"_RcppCWB_cqp_list_corpora", (DL_FUNC) &_RcppCWB_cqp_list_corpora, 0},
     {"_RcppCWB_cqp_query", (DL_FUNC) &_RcppCWB_cqp_query, 3},
-    {"_RcppCWB_cqp_subcorpus_query", (DL_FUNC) &_RcppCWB_cqp_subcorpus_query, 3},
     {"_RcppCWB_cqp_subcorpus_size", (DL_FUNC) &_RcppCWB_cqp_subcorpus_size, 1},
     {"_RcppCWB_cqp_list_subcorpora", (DL_FUNC) &_RcppCWB_cqp_list_subcorpora, 1},
     {"_RcppCWB_cqp_dump_subcorpus", (DL_FUNC) &_RcppCWB_cqp_dump_subcorpus, 1},
@@ -884,6 +883,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_RcppCWB_cqp_drop_subcorpus", (DL_FUNC) &_RcppCWB_cqp_drop_subcorpus, 1},
     {"_RcppCWB_check_corpus", (DL_FUNC) &_RcppCWB_check_corpus, 1},
     {"_RcppCWB_cqp_load_corpus", (DL_FUNC) &_RcppCWB_cqp_load_corpus, 2},
+    {"_RcppCWB_cqp_subcorpus", (DL_FUNC) &_RcppCWB_cqp_subcorpus, 3},
     {"_RcppCWB_cwb_makeall", (DL_FUNC) &_RcppCWB_cwb_makeall, 3},
     {"_RcppCWB_cwb_huffcode", (DL_FUNC) &_RcppCWB_cwb_huffcode, 3},
     {"_RcppCWB_cwb_compress_rdx", (DL_FUNC) &_RcppCWB_cwb_compress_rdx, 3},